feat(TUP-31487): TCK Basic component setup (#4402)
* feat(TUP-31487): add couchbase feature

* feat(TUP-31487): do not remove car after installation

* feat(TUP-31487): install tck in backgroud

* feat(TUP-31487): add slf4j

* feat(TUP-31487): install all components

* feat(TUP-31487): process snapshot

* feat(TUP-31487): move couchbase to bd

* feat(TUP-31487): move couchbase to bd

* feat(TUP-31487): remove couchbase from tos

* feat(TUP-31487): merge master

* feat(TUP-31487): remove all tck from tos

* feat(TUP-31487): support install tcompv0

* feat(TUP-31487): get GAV

* feat(TUP-31487): add back tcompv0

* feat(TUP-31487): install cars without updating config.ini

+++ main/plugins/org.talend.updates.runtime/src/main/java/org/talend/updates/runtime/model/ExtraFeature.java
@@ -191,6 +191,10 @@ public interface ExtraFeature extends Comparable<Object> {
     }
 
     default public void syncComponentsToInstalledFolder(IProgressMonitor progress, File downloadedCompFile) {
+        syncComponentsToInstalledFolder(progress, downloadedCompFile, true);
+    }
+
+    public default void syncComponentsToInstalledFolder(IProgressMonitor progress, File downloadedCompFile, boolean deleteCompFile) {
         // try to move install success to installed folder
         if (progress == null) {
             progress = new NullProgressMonitor();
@@ -203,7 +207,9 @@ public interface ExtraFeature extends Comparable<Object> {
             final File installedComponentFile = new File(installedComponentFolder, downloadedCompFile.getName());
             if (!installedComponentFile.equals(downloadedCompFile)) { // not in same folder
                 FilesUtils.copyFile(downloadedCompFile, installedComponentFile);
-                downloadedCompFile.delete();
+                if (deleteCompFile) {
+                    downloadedCompFile.delete();
+                }
                 progress.worked(1);
             }
 
@@ -212,7 +218,7 @@ public interface ExtraFeature extends Comparable<Object> {
             ExceptionHandler.process(e);
         }
     }
-
+    
     default void shareComponent(IProgressMonitor progress, File installedCompFile) {
         if (!isShareEnable()) {
             return;
